backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "openai>=1.12.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "openai-agents>=0.1.0",
    "slowapi>=0.1.9",
]
